+++ src/parser/enrichers/class/barbarian/SpiritShield.mjs
@@ -1,7 +1,7 @@
 /* eslint-disable class-methods-use-this */
 import DDBEnricherData from "../../data/DDBEnricherData.mjs";
 
-export default class FlashOfGenius extends DDBEnricherData {
+export default class SpiritShield extends DDBEnricherData {
 
   get type() {
     return "utility";